cleanup sdk usage
