+++ crates/sui-graphql-rpc/Cargo.toml
@@ -50,7 +50,6 @@ bcs.workspace = true
 move-transactional-test-runner.workspace = true
 simulacrum.workspace = true  # todo: cleanup test only deps
 sui-transactional-test-runner.workspace = true # todo: cleanup test only deps
-sui-sdk.workspace = true
 sui-json-rpc.workspace = true
 sui-json-rpc-types.workspace = true
 sui-indexer.workspace = true
