+++ crates/sui-graphql-rpc/src/context_data/mod.rs
@@ -4,6 +4,5 @@
 pub(crate) mod db_data_provider;
 pub mod db_query_cost;
 pub(crate) mod package_cache;
-pub(crate) mod sui_sdk_data_provider;
 
 pub const DEFAULT_PAGE_SIZE: u64 = 10;
